requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
python-dotenv
sentence-transformers
transformers
pillow
